int -> long

+++ lab_2_1_2/main.c
@@ -1,69 +1,66 @@
 #include <stdio.h>
 #include <math.h>
 
-int n_input(void); // Input number of elements
-int arr_in(int *a, int n); // Input array elements
-int is_prime(int x); // Check if number is prime
-int prime_array(int *in_array, int *out_array, int n); //Form new array with primes
-int arr_print(int *array, int n); // Print an array
+long n_input(void);                                        // Input number of elements
+long arr_in(long *a, long n);                              // Input array elements
+long is_prime(long x);                                     // Check if number is prime
+long prime_array(long *in_array, long *out_array, long n); //Form new array with primes
+long arr_print(long *array, long n);                       // printarr_print an array
 
-int main(void)
+long main(void)
 {
     int error = 1;
-    
-    int ni = n_input(); // Arraysize variable
+
+    long ni = n_input(); // Arraysize variable
     if (ni > 0)
     {
-        int inarr[10]; // Initial array
+        long inarr[10]; // Initial array
 
         if (arr_in(inarr, ni) == 0)
         {
-            int outarr[10]; // Output array
-            int no = prime_array(inarr, outarr, ni); // Number of elems in output array
+            long outarr[10];                          // Output array
+            long no = prime_array(inarr, outarr, ni); // Number of elems in output array
             if (no > 0)
                 error = arr_print(outarr, no);
             else
                 printf("There's no prime elements");
         }
     }
-    
+
     return error;
 }
 
-
-int arr_print(int *array, int n)
+long arr_print(long *array, long n)
 {
     printf("Out:\n");
 
-    for (int i = 0; i < n; i++)
-        printf("%d ", array[i]);
-    
+    for (long i = 0; i < n; i++)
+        printf("%ld ", array[i]);
+
     return 0;
 }
 
-
-int prime_array(int *in_array, int *out_array, int n)
+long prime_array(long *in_array, long *out_array, long n)
 {
-    int j = 0;
-    for (int i = 0; i < n; i++)
+    long j = 0;
+    for (long i = 0; i < n; i++)
     {
         if (is_prime(in_array[i]) == 0)
         {
             out_array[j] = in_array[i];
             j++;
-        }    
+        }
     }
 
     return j;
 }
 
-
-int is_prime(int x)
+long is_prime(long x)
 {
-    int state = 0;
+    long state = 0;
     if (x > 0)
     {
-        for (int i = 2; i <= sqrt(x); i++)
+        for (long i = 2; i <= sqrt(x); i++)
             if (x % i == 0)
             {
                 state = 1;
@@ -74,18 +71,17 @@ int is_prime(int x)
     {
         state = 1;
     }
-    
+
     return state;
 }
 
-
-int arr_in(int *a, int n)
-{   
+long arr_in(long *a, long n)
+{
     printf("Input array elements:\n");
 
-    for (int i = 0; i < n; i++)
-    {   
-        if (scanf("%d", &a[i]) != 1)
+    for (long i = 0; i < n; i++)
+    {
+        if (scanf("%ld", &a[i]) != 1)
         {
             printf("Incorrect data");
             return 1;
@@ -95,12 +91,11 @@ int arr_in(int *a, int n)
     return 0;
 }
 
-
-int n_input(void)
+long n_input(void)
 {
-    int n;
+    long n;
     printf("Input amount of elements: ");
-    if (scanf("%d", &n) != 1 || n > 10 || n < 1)
+    if (scanf("%ld", &n) != 1 || n > 10 || n < 1)
     {
         printf("Icorrect data");
         return -1;
